Initial Commit
This is untested code and just the rough framework for the start of the
Doge Faucet 2.0 project. Be aware this is NOT complete and will more
than likely fail to function. When the codebase is at a mature state I
will merge the development branch into main and make an announcement.

Thank you,
John Annis